pyproject.toml
[project]
name = "utenagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi[standard]>=0.135.1",
    "python-jose[cryptography]>=3.3.0",
    "langchain>=1.2.13",
    "langchain-text-splitters>=1.1.1",
    "passlib>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.10",
    "recursivecharactertextsplitter>=0.1.0",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
]
